[core] move config group into its own file
`Config` is a very simple key-value-pair storage-and-retrieval system, `ConfigGroup` has PHP reflection magic used to determine key-value-pairs; they are related but not the same
